[Refactor] 코드 정리 (2023/09/18)
