+++ src/components/button/ProfileImage.jsx
@@ -21,7 +21,8 @@ function ProfileImage({
       <img
         src={imageName[1] === '' ? lionHeadLogo : imageURL}
         alt={`${nickname}의 프로필 이미지`}
-        className={`w-[${size}px] h-[${size}px] min-w-[${size}px] min-h-[${size}px] rounded-full border-2 border-lionly-gray-4 bg-cover bg-no-repeat shadow-lg`}
+        className={`min-w-[${size}px] min-h-[${size}px] rounded-full border-2 border-lionly-gray-4 bg-cover bg-no-repeat shadow-lg`}
+        style={{ width: `${size}px`, height: `${size}px` }}
       />
 
       {handleInputClick ? (
