+++ src/hooks/index.js
@@ -1,5 +1,6 @@
 export { default as useChannel } from './useChannel';
 export { default as useContent } from './useContent';
+export { default as useComments } from './useComments';
 export { default as useInfiniteFeed } from './useInfiniteFeed';
 export { default as useObserveScroll } from './useObserveScroll';
 export { default as useScroll } from './useScroll';
